AT-465 Clone SDI SSH key distribution project & add our keys to it.
The SSH keys have been directly copied from the resources/users/*/.ssh folders. I also added each user's name to the end of the ssh-key line in the vars file. This is used to create users on the hosts.

I would like to update this usage so that Ansible instead loads the keys from the resources folder, but this will take a bit more time and research. So for a initial implementation, I decided to just copy the keys.

The CI/CD file has also been updated to reference the CI/CD file in the cloned project so that it runs the ansible linting job.
However, the ansible-lint job does not work at the moment.
1. `ansible-lint` produces the following error on Gitlab: `ERROR: Ansible requires the locale encoding to be UTF-8; Detected None.`
   - I tried a few different python images and they all had the same issue.
2. Running `ansible-lint` locally on the project produces 36 failures.