[AT-628] - Finer grained control on how kube context is set.
This change adds an 'admin' kubectl setup script which sets the kube context to an admin context if the users are in an admin group.
The admin group is currently only setup for atlas.
Arguably, this can be simplified to just check if the team=atlas and then have specific user checks.
But I'd like to keep some of the other changes anyway, because they make dealing with groups a bit easier.