[AT-1032] - Switch to using numeric values for file/dir modes.
The symbolic values were not working as expected.
If the 'other' value was not specified, it seems to have either not updated the value, or allowed reads.
This resulted in other teams being able to read a team's shared dir.
I also restricted read access to the ssh config file as well.