[AT-1857] - Fix image tagging
The main image was being tagged with 'base' as its name, which was resulting in both images sharing the same name in the registry.
Additionally, in nexus, the image directory is the only identifier used when publishing the image. Therefore, we need to add the 'ska-mid-itf' prefix to the image names.